requirements.txt
boto3
seldon-core
grpcio
onnxruntime
numpy
Pillow